Add files via upload
Added 6)c) to code. 6)c) uses `xor` between each element in the piles list and gives a result. If the result is not 0 then the AI will make a move that will leave the user with a list of piles that will xor sum to 0 leaving the AI in a winning position. However, if the xor sum is 0 on the AI's go then the AI will use stratB for their next move.